requirements.txt
pyTelegramBotAPI==4.29.1
loguru==0.7.3
python-dotenv==1.1.1
